% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_loglik.R
\name{CompLogLik}
\alias{CompLogLik}
\title{Create Component Log-likelihood object}
\description{
Create Component Log-likelihood object

Create Component Log-likelihood object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompLogLik-new}{\code{CompLogLik$new()}}
\item \href{#method-CompLogLik-mark_stale}{\code{CompLogLik$mark_stale()}}
\item \href{#method-CompLogLik-set_param}{\code{CompLogLik$set_param()}}
\item \href{#method-CompLogLik-get_value}{\code{CompLogLik$get_value()}}
\item \href{#method-CompLogLik-revert_value}{\code{CompLogLik$revert_value()}}
\item \href{#method-CompLogLik-cache_value}{\code{CompLogLik$cache_value()}}
\item \href{#method-CompLogLik-clone}{\code{CompLogLik$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompLogLik-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompLogLik-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompLogLik$new(param_names, fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Mark Component Log-likelihood as Out-of-date (stale)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompLogLik-mark_stale"></a>}}
\if{latex}{\out{\hypertarget{method-CompLogLik-mark_stale}{}}}
\subsection{Method \code{mark_stale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompLogLik$mark_stale()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Set Parameter and Update Component Log-Likelihood
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompLogLik-set_param"></a>}}
\if{latex}{\out{\hypertarget{method-CompLogLik-set_param}{}}}
\subsection{Method \code{set_param()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompLogLik$set_param(param, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Get Component Log-likelihood Value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompLogLik-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-CompLogLik-get_value}{}}}
\subsection{Method \code{get_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompLogLik$get_value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Revert Value to Previous Value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompLogLik-revert_value"></a>}}
\if{latex}{\out{\hypertarget{method-CompLogLik-revert_value}{}}}
\subsection{Method \code{revert_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompLogLik$revert_value()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Cache Value of Component Log-likelihood for Future Retrieval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompLogLik-cache_value"></a>}}
\if{latex}{\out{\hypertarget{method-CompLogLik-cache_value}{}}}
\subsection{Method \code{cache_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompLogLik$cache_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompLogLik-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompLogLik-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompLogLik$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
