% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{MCMC}
\alias{MCMC}
\title{R6 Class for MCMC sampler}
\description{
R6 Class for MCMC sampler

R6 Class for MCMC sampler
}
\details{
An MCMC class is designed to abstract the sampling algorithm infrastructure
from the model definition, and enable faster iteration on model development
and selection, minimizing the need for changing MCMC code for different models.
Create an MCMC Sampler Object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MCMC-new}{\code{MCMC$new()}}
\item \href{#method-MCMC-calc_gamma1}{\code{MCMC$calc_gamma1()}}
\item \href{#method-MCMC-set_prop}{\code{MCMC$set_prop()}}
\item \href{#method-MCMC-eval_prop_log_lik}{\code{MCMC$eval_prop_log_lik()}}
\item \href{#method-MCMC-make_proposal}{\code{MCMC$make_proposal()}}
\item \href{#method-MCMC-full_cond}{\code{MCMC$full_cond()}}
\item \href{#method-MCMC-update_param}{\code{MCMC$update_param()}}
\item \href{#method-MCMC-gibbs_update}{\code{MCMC$gibbs_update()}}
\item \href{#method-MCMC-mh_update}{\code{MCMC$mh_update()}}
\item \href{#method-MCMC-store_sample}{\code{MCMC$store_sample()}}
\item \href{#method-MCMC-set_n_mcmc}{\code{MCMC$set_n_mcmc()}}
\item \href{#method-MCMC-allocate_sample_storage}{\code{MCMC$allocate_sample_storage()}}
\item \href{#method-MCMC-add_n_mcmc}{\code{MCMC$add_n_mcmc()}}
\item \href{#method-MCMC-add_sample_storage}{\code{MCMC$add_sample_storage()}}
\item \href{#method-MCMC-run_mcmc}{\code{MCMC$run_mcmc()}}
\item \href{#method-MCMC-run_mcmc_part}{\code{MCMC$run_mcmc_part()}}
\item \href{#method-MCMC-clear_cache}{\code{MCMC$clear_cache()}}
\item \href{#method-MCMC-write_to_file}{\code{MCMC$write_to_file()}}
\item \href{#method-MCMC-prepend_samples}{\code{MCMC$prepend_samples()}}
\item \href{#method-MCMC-prepend_stored_samples}{\code{MCMC$prepend_stored_samples()}}
\item \href{#method-MCMC-post_summary}{\code{MCMC$post_summary()}}
\item \href{#method-MCMC-plot_post_histogram}{\code{MCMC$plot_post_histogram()}}
\item \href{#method-MCMC-plot_post_trace}{\code{MCMC$plot_post_trace()}}
\item \href{#method-MCMC-clone}{\code{MCMC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-new"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$new(
  data,
  init,
  priors,
  proposals,
  log_lik,
  n_mcmc,
  n_adapt,
  n_burnin,
  thin_int,
  cache_freq,
  cache_dir,
  combine_cached,
  const,
  show_progress
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(named list) of all data elements used in likelihood calculations}

\item{\code{init}}{(named list) of initial values for all parameters for which to draw
posterior samples. Parameter names listed here will be available in
\code{param} argument passed to user-defined likelihood functions. See:
\code{log_lik} argument and \code{LogLik} class.}

\item{\code{priors}}{(named list)}

\item{\code{proposals}}{(named list)}

\item{\code{log_lik}}{(LogLik) object}

\item{\code{n_mcmc}}{(integer)}

\item{\code{n_adapt}}{(integer)}

\item{\code{n_burnin}}{(integer)}

\item{\code{thin_int}}{(integer)}

\item{\code{cache_freq}}{(integer)}

\item{\code{cache_dir}}{(string)}

\item{\code{combine_cached}}{(logical)}

\item{\code{const}}{(named list)}

\item{\code{show_progress}}{(logical)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `MCMC` sampler object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-calc_gamma1"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-calc_gamma1}{}}}
\subsection{Method \code{calc_gamma1()}}{
Calculate gamma1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$calc_gamma1()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-set_prop"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-set_prop}{}}}
\subsection{Method \code{set_prop()}}{
Set Proposal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$set_prop(param_name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-eval_prop_log_lik"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-eval_prop_log_lik}{}}}
\subsection{Method \code{eval_prop_log_lik()}}{
Evaluate Proposal Log-likelihood
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$eval_prop_log_lik(param_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param_names}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-make_proposal"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-make_proposal}{}}}
\subsection{Method \code{make_proposal()}}{
Make Proposal
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$make_proposal(param_name, block_level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{block_level}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-full_cond"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-full_cond}{}}}
\subsection{Method \code{full_cond()}}{
Calculate Full Conditional
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$full_cond(param_name, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-update_param"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-update_param}{}}}
\subsection{Method \code{update_param()}}{
Update Parameter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$update_param(param_name, block_level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{block_level}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-gibbs_update"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-gibbs_update}{}}}
\subsection{Method \code{gibbs_update()}}{
Gibbs Update
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$gibbs_update(param_name, block_level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{block_level}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-mh_update"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-mh_update}{}}}
\subsection{Method \code{mh_update()}}{
Metropolis-Hastings Update
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$mh_update(param_name, block_level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{block_level}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-store_sample"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-store_sample}{}}}
\subsection{Method \code{store_sample()}}{
Store Sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$store_sample(param_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param_name}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-set_n_mcmc"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-set_n_mcmc}{}}}
\subsection{Method \code{set_n_mcmc()}}{
Set Number of MCMC Samples to Draw
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$set_n_mcmc(n_mcmc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_mcmc}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-allocate_sample_storage"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-allocate_sample_storage}{}}}
\subsection{Method \code{allocate_sample_storage()}}{
Allocate Sample Storage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$allocate_sample_storage()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-add_n_mcmc"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-add_n_mcmc}{}}}
\subsection{Method \code{add_n_mcmc()}}{
Add n MCMC Samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$add_n_mcmc(n_add_mcmc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_add_mcmc}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-add_sample_storage"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-add_sample_storage}{}}}
\subsection{Method \code{add_sample_storage()}}{
Add Sample Storage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$add_sample_storage(n_add_mcmc)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_add_mcmc}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-run_mcmc"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-run_mcmc}{}}}
\subsection{Method \code{run_mcmc()}}{
Run MCMC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$run_mcmc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-run_mcmc_part"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-run_mcmc_part}{}}}
\subsection{Method \code{run_mcmc_part()}}{
Run MCMC Part
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$run_mcmc_part()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-clear_cache"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-clear_cache}{}}}
\subsection{Method \code{clear_cache()}}{
Clear Cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$clear_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-write_to_file"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-write_to_file}{}}}
\subsection{Method \code{write_to_file()}}{
Write Samples to File
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$write_to_file(file_path, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-prepend_samples"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-prepend_samples}{}}}
\subsection{Method \code{prepend_samples()}}{
Prepennd Samples to Sampler Storage
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$prepend_samples(samples)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samples}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-prepend_stored_samples"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-prepend_stored_samples}{}}}
\subsection{Method \code{prepend_stored_samples()}}{
Prepend Stored Samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$prepend_stored_samples(file_paths)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_paths}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-post_summary"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-post_summary}{}}}
\subsection{Method \code{post_summary()}}{
Summarize Posterior Samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$post_summary(probs = c(0.025, 0.975))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{probs}}{(vector of Real numbers between [0,1])}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(named list of tibbles) for each parameter summarizing posterior
samples.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-plot_post_histogram"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-plot_post_histogram}{}}}
\subsection{Method \code{plot_post_histogram()}}{
Plot Posterior Histograms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$plot_post_histogram()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-plot_post_trace"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-plot_post_trace}{}}}
\subsection{Method \code{plot_post_trace()}}{
Plot Posterior Traceplots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$plot_post_trace()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MCMC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MCMC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MCMC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
