% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{zero_pad_strd}
\alias{zero_pad_strd}
\title{Prepend Zeros to Integer and Convert to String}
\usage{
zero_pad_strd(x, width = 10)
}
\arguments{
\item{x}{(integer) to be converted to string. If x has more than \code{width}
digits, no 0s will be prepended.}

\item{width}{(integer) desired fixed width of string (default = 10 characters)}
}
\description{
Prepend Zeros to Integer and Convert to String
}
